third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/html/semantics/forms/textfieldselection/selection-not-application-textarea-expected.txt

This is a testharness.js-based test.
FAIL text field selection for the input textarea assert_equals: updated selectionStart expected 1 but got 0
Harness: the test ran to completion.
